Moving GC into image manager
